% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{ahmisc-package}
\alias{ahmisc-package}
\title{ahmisc}
\description{
Miscellaneous functions and templates
}
\section{Rmarkdown templates}{

Templates "notebook" (general programming), "report" (share results of a
project) and "consulting" (generate PDF output for consulting cases).
}

